Tests: flaky tests 3324 - Added Repeated trait
